sms_sender/requirements.txt
uwsgi
gunicorn
django
django_uwsgi
pycryptodome
huawei_lte_api
djangorestframework 
markdown
django-filter